+++ tests/test_logout.py
@@ -22,9 +22,9 @@ from django.test import Client
 from bastion import connections as connections_module
 from bastion.connections import Connection
 from bastion.protocols.oidc.transaction import build_end_session_url
+from bastion.testing.provider import FakeIdP
+from bastion.testing.transport import FakeTransport
 from bastion.views import SESSION_CONNECTION_KEY, SESSION_ID_TOKEN_KEY
-from tests.idp.provider import FakeIdP
-from tests.idp.transport import FakeTransport
 from tests.test_login_flow import login, make_connection, only_connections
 
 pytestmark = pytest.mark.django_db